rename check-litellm exporter
