+++ python-sdks/respan-exporter-litellm/.env.example
@@ -1,3 +1,3 @@
-KEYWORDSAI_BASE_URL=http://localhost:8000/api
-KEYWORDSAI_API_KEY=YOUR_KEYWORDSAI_API_KEY
+RESPAN_BASE_URL=http://localhost:8000/api
+RESPAN_API_KEY=YOUR_RESPAN_API_KEY
 OPENAI_API_KEY=YOUR_OAI_API_KEY
